clippy: Fix a bunch of warnings in `script`
This is just a portion of the errors that are remaining to be fixed.

+++ components/script/dom/gpucompilationmessage.rs
@@ -69,10 +69,10 @@ impl GPUCompilationMessage {
             global,
             info.message.into(),
             GPUCompilationMessageType::Error,
-            info.line_number as u64,
-            info.line_pos as u64,
-            info.offset as u64,
-            info.length as u64,
+            info.line_number,
+            info.line_pos,
+            info.offset,
+            info.length,
         )
     }
 }
